#1 중복 contract 디렉토리 하나로 통일, 기본 도메인 repository들 domain 하위 repository로 파일 위치 변경
