+++ src/main/java/com/example/api/board/service/CategoryService.java
@@ -1,11 +1,10 @@
 package com.example.api.board.service;
 
 import com.example.api.board.controller.domain.CategoryDTO;
-import com.example.api.board.repository.CategoryRepository;
+import com.example.api.domain.repository.CategoryRepository;
 import com.example.api.domain.Category;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
