+++ src/main/java/com/example/api/contracts/controller/ContractController.java
@@ -1,5 +1,6 @@
 package com.example.api.contracts.controller;
 
+import com.example.api.contracts.dto.ContractDTO;
 import com.example.api.contracts.ContractService;
 import com.example.api.contracts.dto.AcceptContractCommand;
 import com.example.api.contracts.dto.AcceptSuggestCommand;
@@ -59,5 +60,11 @@ class ContractController {
         final AcceptContractCommand acceptContractCommand = new AcceptContractCommand(contractId);
         contractService.acceptContract(acceptContractCommand);
         return ResponseEntity.ok(null);
+    }
+
+    @GetMapping("/api/v1/contract/{contractId}/status")
+    public ResponseEntity<ContractDTO> getContractInfo(@PathVariable("contractId") long contractId) {
+        ContractDTO contractDTO = contractService.getContractInfo(contractId);
+        return ResponseEntity.ok(contractDTO);
     }
 }
