+++ src/main/java/com/example/api/contracts/dto/EmployeeInfoDTO.java
@@ -1,4 +1,4 @@
-package com.example.api.contract.controller.domain;
+package com.example.api.contracts.dto;
 
 import lombok.*;
