+++ src/test/java/com/example/api/global/BaseIntegrationTest.java
@@ -4,9 +4,8 @@ import com.example.api.board.controller.domain.CategoryDTO;
 import com.example.api.board.controller.domain.ExternalCareerDTO;
 import com.example.api.board.controller.domain.InnerCareerDTO;
 import com.example.api.board.controller.domain.PossibleBoardDTO;
-import com.example.api.board.repository.*;
 import com.example.api.domain.*;
-import jakarta.transaction.Transactional;
+import com.example.api.domain.repository.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
